anoubisctl: Support monitor without sending event replies
anoubisctl monitor beantwortet nicht mehr ask-policies wenn kein
error=X ueber die Kommandozeile mitgegeben wird.
	 -src/anoubis/src/anoubisctl/main.c
	 ask policies werden nur noch beantwortet wenn error >= 0
	 -test/anoubisctl/XenAnoubisCtl.pm
	 und test/anoubisctl/XenAnoubisCtl/tests/1010
	 Test ueberprueft ob dem auch so ist
	 -test/sfs/DaemonPolicy.pm
	 und test/sfs/DaemonPolicy/tests/2012
	 Anpassung eines alten Tests an das neue Verhalten

Aegis-Change: anoubis.10.970.C341 Delta 150
Bug: 1250
BLE

+++ src/anoubisctl/main.c
@@ -797,8 +797,6 @@ static int monitor(int argc, char **argv)
 	}
 	if (maxcount < 0)
 		maxcount = 0;
-	if (defaulterror < 0)
-		defaulterror = 0;
 
 	error = create_channel();
 	if (error) {
@@ -847,14 +845,18 @@ static int monitor(int argc, char **argv)
 				anoubis_msg_free(m);
 				continue;
 			}
-			error = anoubis_client_notifyreply(client,
-			    m->u.notify->token, defaulterror, delegate);
-			anoubis_msg_free(m);
-			if (error < 0) {
-				fprintf(stderr, "notify reply failed "
-				    "with code %d", -error);
-				destroy_channel();
-				return 5;
+			if (defaulterror >= 0) {
+				error = anoubis_client_notifyreply(client,
+				    m->u.notify->token, defaulterror, delegate);
+				anoubis_msg_free(m);
+				if (error < 0) {
+					fprintf(stderr, "notify reply failed "
+					    "with code %d", -error);
+					destroy_channel();
+					return 5;
+				}
+			} else {
+				anoubis_msg_free(m);
 			}
 		}
 		if (maxcount > 0 && count >= maxcount)
